processes and dumps all linux major tags now, very slow

+++ hancock/main.py
@@ -85,38 +85,44 @@ def get_files_for_diff(tag1, tag2, repo_path):
     files = [file for file in files if file.endswith('.c') or file.endswith('.h')]
     return files
 
+def dump_functions_for_tag(tag, repo_path, output_path):
+    print(f"Processing {tag}")
+    output_path = os.path.join(output_path, f"{tag}.json")
+    if os.path.exists(output_path):
+        print(f"Skipping {tag} because {output_path} already exists")
+        return
+    output = subprocess.run(['git', 'checkout', '-f', tag], cwd=repo_path)
+    files = get_files_for_tag(tag, repo_path)
+    functions = process_files(files, repo_path)
+    with open(output_path, 'w') as f:
+        json.dump(functions, f, indent=2)
+    print(f"Wrote {len(functions)} functions to {output_path}")
+
 def main():
     # Get command line args
     parser = argparse.ArgumentParser()
     parser.add_argument('--repo', type=str, default='artifacts/linux', help='Path to the repository')
-    parser.add_argument('--output', type=str, default='blobs/linux_functions.json', help='Path to the output file')
+    parser.add_argument('--output', type=str, default='blobs/linux_functions', help='Path to the output file')
     args = parser.parse_args()
+
+    # Make the output directory
+    os.makedirs(args.output, exist_ok=True)
 
     # Get the tags
     tags = get_indexed_tags(args.repo)
-
-    #Checkout the first tag
-    output = subprocess.run(['git', 'checkout', '-f', tags[0][1]], cwd=args.repo)
-
-    print(output)
+    
+    print(f"Processing {len(tags)} tags")
 
     # Get the functions all the files with the first tag 
-    files = get_files_for_tag(tags[0][1], args.repo)
-    print(f"Found {len(files)} files in {tags[0][1]}")
+    for tag in tags:
+        dump_functions_for_tag(tag[1], args.repo, args.output)
 
-    # parallel extraction of functions from all files
-    functions = process_files(files, args.repo)
-    print(f"Found {len(functions)} functions in {len(files)} files")
-    # saved the functions to a file
-    with open(args.output, 'w') as f:
-        json.dump(functions, f, indent=2)
-
-    # get list of files that changed for each tag 
-    for i in range(1, len(tags)):
-        tag1 = tags[i-1][1]
-        tag2 = tags[i][1]
-        files = get_files_for_diff(tag1, tag2, args.repo)
-        print(f"Comparing {tag1} to {tag2}, {len(files)} files changed")
+    # # get list of files that changed for each tag 
+    # for i in range(1, len(tags)):
+    #     tag1 = tags[i-1][1]
+    #     tag2 = tags[i][1]
+    #     files = get_files_for_diff(tag1, tag2, args.repo)
+    #     print(f"Comparing {tag1} to {tag2}, {len(files)} files changed")
